refactor: create TypedSearchAggregationQueryTransformer & move aggregation query logic
