+++ search/search-client-elasticsearch/src/main/java/io/camunda/search/es/transformers/aggregator/SearchAggregationResultTransformer.java
@@ -15,8 +15,8 @@ import co.elastic.clients.elasticsearch._types.aggregations.SingleBucketAggregat
 import co.elastic.clients.elasticsearch._types.aggregations.StringTermsBucket;
 import co.elastic.clients.elasticsearch.core.SearchResponse;
 import io.camunda.search.clients.transformers.SearchTransfomer;
-import io.camunda.search.clients.transformers.aggregate.SearchAggregationResult;
-import io.camunda.search.clients.transformers.aggregate.SearchAggregationResult.Builder;
+import io.camunda.search.clients.transformers.aggregation.SearchAggregationResult;
+import io.camunda.search.clients.transformers.aggregation.SearchAggregationResult.Builder;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
